web-scraper/requirements.txt
argparse
playwright
beautifulsoup4
pandas
tqdm
psycopg2-binary
pdfplumber
requests
